+++ src/nbb/core.cljs
@@ -611,10 +611,13 @@
                                'type->str (sci/copy-var type->str core-ns)
                                'Keyword (sci/copy-var Keyword core-ns)
                                'Symbol (sci/copy-var Symbol core-ns)
-                               'PersistentVector (sci/copy-var PersistentVector core-ns)
+                               'PersistentVector PersistentVector
                                'IFn (sci/copy-var IFn core-ns)
                                'swap-vals! (sci/copy-var swap-vals! core-ns)
-                               'reset-vals! (sci/copy-var reset-vals! core-ns)}
+                               'reset-vals! (sci/copy-var reset-vals! core-ns)
+                               'PersistentQueue (let [x PersistentQueue]
+                                                  (gobj/set x "EMPTY" cljs.core/PersistentQueue.EMPTY)
+                                                  x)}
                 'cljs.reader {'read-string (sci/copy-var edn/read-string (sci/create-ns 'cljs.reader))}
                 'clojure.main {'repl-requires (sci/copy-var
                                                repl-requires